moved main page to a smaller component
