+++ src/types/types.ts
@@ -1,5 +1,8 @@
+import { Key, JSX } from "react";
+
 // src/types/types.ts
 export interface Section {
+    map(arg0: (section: Section, index: Key | null | undefined) => JSX.Element | undefined): import("react").ReactNode;
     id: number;
     blog_id: number;
     section_type_id: number;
